Restructured functions and have started working on json parsing for sample dataset.
