+++ internal/find/find_one.go
@@ -1,0 +1,5 @@
+// https://pkg.go.dev/go.mongodb.org/mongo-driver/mongo#Collection.FindOne
+
+package find
+
+func FindOne() {}
