+++ samples/dataset_films.json
@@ -1,0 +1,28 @@
+{
+  "films": [
+    {
+      "name": "",
+      "year": "",
+      "genre": "",
+      "language": ""
+    },
+    {
+      "name": "",
+      "year": "",
+      "genre": "",
+      "language": ""
+    },
+    {
+      "name": "",
+      "year": "",
+      "genre": "",
+      "language": ""
+    },
+    {
+      "name": "",
+      "year": "",
+      "genre": "",
+      "language": ""
+    }
+  ]
+}
